Extend GeophiresXTestCase from BaseTestCase
